quest: Allow ItemId enum for quest files
Added the ability to provide reward items using the ItemId enum instead
of numeric literals in quest JSON files.

+++ Arrowgene.Ddon.Shared/Model/DropsTable.cs
@@ -18,7 +18,7 @@ public class DropsTable
     {
         Items.Add(new GatheringItem()
         {
-            ItemId = (uint)itemId,
+            ItemId = itemId,
             ItemNum = minAmount,
             MaxItemNum = maxAmount,
             DropChance = dropChance
